extract address
