+++ gb-bus/src/lib.rs
@@ -1,8 +1,10 @@
+pub mod address;
 pub mod address_bus;
 pub mod file_operation;
 pub mod generic;
 
-pub use address_bus::{Address, AddressBus, Area, Error};
+pub use address::{Address, Area};
+pub use address_bus::{AddressBus, Error};
 pub use file_operation::FileOperation;
 
 trait Bus {
